% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseContent.R
\name{compareBaseContent}
\alias{compareBaseContent}
\title{\strong{Compare Base Content}}
\usage{
compareBaseContent(seqList, bases, baseOnly = TRUE, pct = FALSE, desc = TRUE)
}
\arguments{
\item{seqList}{(required): List of Biostring sequences, each for which the content is to be calculated. It is recommended that this list be named for readability of output. Otherwise, the output will provide the corresponding index occupied by the sequence in the original list that was provided alongside its base content calculation.}

\item{bases}{(required): Character or vector of characters for bases that are to be used for calculating their content proportion in the sequence provided. If a vector of characters is provided, this function finds the proportion of the sequence that is any one of the bases (i.e. necessarily in a contiguous manner).}

\item{baseOnly}{(default TRUE): boolean for if the user wants to calculate the content on using the base pair characters (ex. A, T, C, G for DNA) in the sequence, or from all characters in the sequence.}

\item{pct}{(default FALSE): boolean for if the user wishes to return the output as a percentage (TRUE) or proportion (FALSE, default).}

\item{desc}{(default TRUE): boolean for if the user wishes to return the output in descending order of base content (TRUE, default) or in ascending order (FALSE).}
}
\value{
Named vector of proportions or percentages for base content of each provided sequence, in either descending or ascending order of base content.
}
\description{
Computes the content for each sequence in a list of sequences for either a base or a vector of individual bases, and then outputs them in either ascending or descending order of content.
}
