% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternContent.R
\name{getPatternContent}
\alias{getPatternContent}
\title{\strong{Get Pattern Content}}
\usage{
getPatternContent(seq, pattern, baseOnly = TRUE, ratio = FALSE)
}
\arguments{
\item{seq}{(required): Biostring sequence in which the content is to be calculated.}

\item{pattern}{(required): String for a contiguous pattern of bases that are to be used for calculating their observed and expected occurrences in the sequence provided.}

\item{baseOnly}{(default TRUE): boolean for if the user wants to calculate the content on using the base pair characters (ex. A, T, C, G for DNA) in the sequence, or from all characters in the sequence.}

\item{ratio}{(default FALSE): boolean for if the user wishes to return the output as a named list for observed and expected counts (FALSE, default) or a ratio of observed over expected (TRUE).}
}
\value{
Unlike getBaseContent, this provides not a proportion or percentage, but either a list (default) or the ratio of observed to expected.
}
\description{
Computes the observed content in a sequence for a given pattern of bases, as well as the expected content for that pattern.
}
