% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowContent.R
\name{windowPatternContent}
\alias{windowPatternContent}
\title{\strong{Window Pattern Content}}
\usage{
windowPatternContent(seq, windowSize, pattern, na.rm = FALSE, log = FALSE)
}
\arguments{
\item{seq}{(required): Biostring sequence that is to be split into windows, in which the base content is to be calculated for each window.}

\item{windowSize}{(required): The size of each window, in base pairs.}

\item{pattern}{(required): String for a contiguous pattern of bases that are to be used for calculating their observed and expected occurrences in each created window from the sequence.}

\item{na.rm}{(default FALSE): boolean for if the user wishes to remove all ratios in the returned vector that are NaN.}

\item{log}{(default FALSE): boolean for if the user wishes to natural log transform the returned vector.}
}
\value{
A vector of observed to expected ratios for base content in each window of the user-specified size for a sequence.
}
\description{
Given an individual sequence, this function splits the provided sequence into windows of a user-specified size and then computes the content in each window for a contiguous pattern of bases.
}
