% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseContent.R
\name{compareBaseContent}
\alias{compareBaseContent}
\title{\strong{Compare Base Content}}
\usage{
compareBaseContent(seqList, bases, baseOnly = TRUE, pct = FALSE)
}
\arguments{
\item{seqList}{(required): BSgenome object or a list of sequences, each for which the content is to be calculated. It is recommended that the elements of this data structure be named for readability of output. Otherwise, the output will provide the corresponding index occupied by the sequence in the original data structure that was provided alongside its base content calculation.}

\item{bases}{(required): Character or vector of characters for bases that are to be used for calculating their content proportion in the sequence provided. If a vector of characters is provided, this function finds the proportion of the sequence that is any one of the bases (i.e. necessarily in a contiguous manner).}

\item{baseOnly}{(default TRUE): boolean for if the user wants to calculate the content on using the base pair characters (ex. A, T, C, G for DNA) in the sequence, or from all characters in the sequence.}

\item{pct}{(default FALSE): boolean for if the user wishes to return the output as a percentage (TRUE) or proportion (FALSE, default).}
}
\value{
Named vector of proportions or percentages for base content of each provided sequence.
}
\description{
This function computes the content for each sequence in a genome for either a base or a vector of individual bases. This allows the user to compare between distinct sequences and identify which of the sequences provided has the highest or lowest base content for a given genome. For instance, a user may be interested in finding which chromosome in a genome has the highest CG content.
}
