% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternContent.R
\name{getPatternContent}
\alias{getPatternContent}
\title{\strong{Get Pattern Content}}
\usage{
getPatternContent(seq, pattern, baseOnly = TRUE)
}
\arguments{
\item{seq}{(required): Biostring sequence, or a StringSet of multiple Biostring sequences for which the content is to be calculated.}

\item{pattern}{(required): String for a contiguous pattern of bases that are to be used for calculating their observed and expected occurrences in the sequence(s) provided.}

\item{baseOnly}{(default TRUE): boolean for if the user wants to calculate the content on using the base pair characters (ex. A, T, C, G for DNA; A, U, C, G for RNA) in the sequence, or from all characters in the sequence.}
}
\value{
Unlike getBaseContent, this provides not proportions or percentages for each sequence, but either a ratio of observed to expected counts.
}
\description{
This function computes the ratio of observed to expected content for a specific contiguous pattern of bases in a sequence or a StringSet of multiple sequences.
}
